Fix gossip quest lists broken after 3776388.

+++ src/game/GossipDef.cpp
@@ -488,7 +488,7 @@ void PlayerMenu::SendQuestGiverQuestList(QEmote eEmote, std::string const& Title
             }
 
             data << uint32(questID);
-            data << qmi.m_qIcon;
+            data << uint32(qmi.m_qIcon);
             data << uint32(pQuest->GetQuestLevel());
             data.append(title, titleLen + 1);
         }
